pyproject.toml
[project]
name = "edge-serverless-box"
version = "0.1.0"
description = "High-fidelity Lambda runtime for edge and local development, powered by AWS Lambda RIE."
requires-python = ">=3.12"
dependencies = [
    # API Gateway compatible server
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    # Docker operations
    "docker>=6.1.3",
    # Scheduler
    "apscheduler>=3.10.4",
    # JWT authentication
    "pyjwt>=2.8.0",
    # Database (ScyllaDB / Cassandra compatible)
    # cassandra-driver is unused in Gateway and removed to shorten builds
    # AWS SDK (S3/DynamoDB compatible)
    "boto3>=1.34.0",
    # Configuration management
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.1",
    "pyyaml>=6.0",
    "watchdog>=6.0.0",
    "jinja2>=3.1",
    "cachetools>=5.3.0",
    "questionary>=2.0.0",
    "grpcio>=1.59.0",
    "grpcio-tools>=1.59.0",
    "protobuf>=4.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "respx>=0.20.2",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "lefthook",  # Git hooks manager
    "cryptography>=41.0.0",  # For SSL certificate generation
    "requests>=2.31.0",      # For tests
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tools"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[project.scripts]
esb = "tools.cli.main:main"

[tool.pytest.ini_options]
testpaths = ["tests", "tools", "services/common/tests", "services/gateway/tests", "services/orchestrator/tests"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::urllib3.exceptions.InsecureRequestWarning",
]
markers = [
    "slow: marks tests as slow (may take several minutes)",
]
